+++ app/Http/Controllers/Backend/VendorProductImageGalleryController.php
@@ -8,6 +8,7 @@ use App\Models\Product;
 use App\Models\ProductImageGallery;
 use App\Traits\ImageUploadTrait;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class VendorProductImageGalleryController extends Controller
 {
@@ -19,7 +20,11 @@ class VendorProductImageGalleryController extends Controller
     {
         // return
         $product = Product::findOrFail($request->product);
-        return $dataTable->render('vendor.products.image-gallery.image',compact('product'));
+        //  Check Product Vendor Owner
+        if ($product->vendor_id !== Auth::user()->vendorprofile->id) {
+            abort(404);
+        }
+        return $dataTable->render('vendor.products.image-gallery.image', compact('product'));
         // return view('vendor.products.image-gallery.image');
     }
 
@@ -29,7 +34,7 @@ class VendorProductImageGalleryController extends Controller
     public function create()
     {
         //
-       
+
     }
 
     /**
@@ -39,22 +44,21 @@ class VendorProductImageGalleryController extends Controller
     {
         //
         // dd($request->all());
-        
+
         $request->validate([
-            'image.*' => ['required','image','max:2048']
+            'image.*' => ['required', 'image', 'max:2048']
         ]);
 
-        $imagePaths = $this->uploadMultiImage($request,'image','uploads');
+        $imagePaths = $this->uploadMultiImage($request, 'image', 'uploads');
 
-        foreach($imagePaths as $path){
-            $productImageGallery =new ProductImageGallery();
+        foreach ($imagePaths as $path) {
+            $productImageGallery = new ProductImageGallery();
             $productImageGallery->image = $path;
             $productImageGallery->product_id = $request->product;
             $productImageGallery->save();
-
         }
 
-        toastr('Uploaded Successfully!','success');
+        toastr('Uploaded Successfully!', 'success');
         return redirect()->back();
     }
 
@@ -88,9 +92,14 @@ class VendorProductImageGalleryController extends Controller
     public function destroy(string $id)
     {
         $product = ProductImageGallery::findOrFail($id);
+        // dd($product->Product);
+        //  Check Product Vendor Owner
+        if ($product->Product->vendor_id !== Auth::user()->vendorprofile->id) {
+            abort(404);
+        }
         $this->deleteImage($product->image);
         $product->delete();
 
-        return response(['status' => 'success','message' => 'Deleted Successfully!']);
+        return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
     }
 }
